+++ portability-core/src/main/java/org/dataportabilityproject/serviceProviders/google/contacts/GoogleContactsService.java
@@ -19,21 +19,14 @@ package org.dataportabilityproject.serviceProviders.google.contacts;
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.services.people.v1.PeopleService;
 import com.google.api.services.people.v1.PeopleService.People.Connections;
-import com.google.api.services.people.v1.model.EmailAddress;
-import com.google.api.services.people.v1.model.FieldMetadata;
 import com.google.api.services.people.v1.model.GetPeopleResponse;
 import com.google.api.services.people.v1.model.ListConnectionsResponse;
-import com.google.api.services.people.v1.model.Name;
 import com.google.api.services.people.v1.model.Person;
 import com.google.api.services.people.v1.model.PersonResponse;
-import com.google.api.services.people.v1.model.PhoneNumber;
 import com.google.common.annotations.VisibleForTesting;
 import ezvcard.VCard;
-import ezvcard.property.Email;
-import ezvcard.property.StructuredName;
-import ezvcard.property.Telephone;
+
 import java.io.IOException;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.dataportabilityproject.cloud.interfaces.JobDataCache;
@@ -56,12 +49,8 @@ public class GoogleContactsService implements Exporter<ContactsModelWrapper>,
 
   @VisibleForTesting
   // List of all fields we want to get from the Google Contacts API
+  // NB: this will have to be updated as we support more fields
   static final String PERSON_FIELDS = "addresses,emailAddresses,names,phoneNumbers";
-
-  @VisibleForTesting
-  static final int PRIMARY_PREF = 1;
-  @VisibleForTesting
-  static final int SECONDARY_PREF = 2;
 
   public GoogleContactsService(Credential credential, JobDataCache jobDataCache) {
     // TODO(olsona): add permissions/scopes!
@@ -77,88 +66,6 @@ public class GoogleContactsService implements Exporter<ContactsModelWrapper>,
     this.jobDataCache = jobDataCache;
   }
 
-  @VisibleForTesting
-  static VCard convertPersonToVCard(Person person) {
-    VCard vCard = new VCard();
-
-    /* Reluctant to set the VCard.Kind value, since a) there aren't that many type options for
-    Google contacts, b) those type options are often wrong, and c) those type options aren't even
-    reliably in the same place.
-    Source: https://developers.google.com/people/api/rest/v1/people#personmetadata
-    */
-
-    if (person.getNames() != null) {
-      convertToVCardNamesAndPopulate(vCard, person.getNames());
-    }
-
-    if (person.getPhoneNumbers() != null) {
-      // VCard API does not support adding multiple telephone numbers at once
-      person.getPhoneNumbers().forEach(n -> vCard.addTelephoneNumber(convertToVCardTelephone(n)));
-    }
-
-    if (person.getEmailAddresses() != null) {
-      // VCard API does not support adding multiple emails at once
-      person.getEmailAddresses().forEach(e -> vCard.addEmail(convertToVCardEmail(e)));
-    }
-
-    return vCard;
-  }
-
-  private static int getPref(FieldMetadata metadata) {
-    return metadata.getPrimary() ? PRIMARY_PREF : SECONDARY_PREF;
-  }
-
-  @VisibleForTesting
-  private static Email convertToVCardEmail(EmailAddress personEmail) {
-    // TODO(olsona): address Email.displayName
-    // TODO(olsona): address Email.formattedType
-    Email email = new Email(personEmail.getValue());
-    email.setPref(getPref(personEmail.getMetadata()));
-
-    return email;
-  }
-
-  @VisibleForTesting
-  private static void convertToVCardNamesAndPopulate(VCard vCard,
-      List<Name> personNames) {
-    LinkedList<StructuredName> alternateVCardNames = new LinkedList<>();
-    for (Name personName : personNames) {
-      if (personName.getMetadata().getPrimary()) {
-        // This is the primary name for the Person, so it should be the primary name in the VCard.
-        vCard.setStructuredName(convertToVCardNameSingle(personName));
-      } else {
-        alternateVCardNames.add(convertToVCardNameSingle(personName));
-      }
-    }
-    if (vCard.getStructuredName() == null) {
-      // No personName was set as primary, so we'll just get the first alternate name
-      vCard.setStructuredName(alternateVCardNames.pop());
-    }
-
-    vCard.setStructuredNameAlt(
-        alternateVCardNames.toArray(new StructuredName[alternateVCardNames.size()]));
-  }
-
-  @VisibleForTesting
-  private static StructuredName convertToVCardNameSingle(Name personName) {
-    StructuredName structuredName = new StructuredName();
-    structuredName.setFamily(personName.getFamilyName());
-    structuredName.setGiven(personName.getGivenName());
-    structuredName.getPrefixes().add(personName.getHonorificPrefix());
-    structuredName.getSuffixes().add(personName.getHonorificSuffix());
-
-    // TODO(olsona): address formatting, structure, phonetics
-    return structuredName;
-  }
-
-  @VisibleForTesting
-  private static Telephone convertToVCardTelephone(PhoneNumber personNumber) {
-    Telephone telephone = new Telephone(personNumber.getValue());
-    telephone.setPref(getPref(personNumber.getMetadata()));
-    return telephone;
-  }
-
-  @Override
   public ContactsModelWrapper export(ExportInformation continuationInformation) throws IOException {
     // Set up connection
     Connections.List connectionsList = peopleService.people().connections()
@@ -188,7 +95,7 @@ public class GoogleContactsService implements Exporter<ContactsModelWrapper>,
 
     // Convert Persons to VCards
     List<VCard> vCards = personResponseList.stream()
-        .map(a -> convertPersonToVCard(a.getPerson()))
+        .map(a -> GoogleContactToVCard.convert(a.getPerson()))
         .collect(Collectors.toList());
 
     // Determine if there's a next page
@@ -203,5 +110,6 @@ public class GoogleContactsService implements Exporter<ContactsModelWrapper>,
   @Override
   public void importItem(ContactsModelWrapper object) throws IOException {
     // TODO(olsona)
+
   }
 }
